+++ pyproject.toml
@@ -92,11 +92,11 @@ omit = [
 ]
 
 [tool.pytest.ini_options]
-testpaths = ["src/scirpy/tests"]
+testpaths = "src/scirpy/tests"
 xfail_strict = true
-addopts = [
-    "--import-mode=importlib",  # allow using test files with same name
-]
+# addopts = [
+#     "--import-mode=importlib",  # allow using test files with same name
+# ]
 markers = [
     "conda: marks a subset of tests to be ran on the Bioconda CI.",
     "extra: marks tests that require extra dependencies."
