+++ src/scirpy/io/_legacy.py
@@ -5,9 +5,9 @@ from typing import List, cast
 
 import awkward as ak
 from anndata import AnnData
-from packaging import version
+import packaging.version
+from importlib.metadata import version
 
-from .. import __version__
 from ..util import DataHandler, _is_na2
 from ._datastructures import AirrCell
 from ._util import _IOLogger
@@ -27,19 +27,16 @@ def upgrade_schema(adata: AnnData) -> AnnData:
 
     # Raise error if already up to date
     if ("airr" in adata.obsm and isinstance(adata.obsm["airr"], ak.Array)) or (
-        version.parse(adata.uns.get("scirpy_version", "v0.0.0"))
-        >= version.parse("0.13.0")
+        packaging.version.parse(adata.uns.get("scirpy_version", "v0.0.0")) >= packaging.version.parse("0.13.0")
     ):
-        raise ValueError(
-            "Your AnnData object seems already up-to-date with scirpy v0.13"
-        )
+        raise ValueError("Your AnnData object seems already up-to-date with scirpy v0.13")
     # Raise error if very old schema
     DataHandler.check_schema_pre_v0_7(adata)
 
     airr_cells = _obs_schema_to_airr_cells(adata)
     tmp_adata = from_airr_cells(airr_cells)
     adata.obsm["airr"] = tmp_adata.obsm["airr"]
-    adata.uns["scirpy_version"] = __version__
+    adata.uns["scirpy_version"] = version("scirpy")
     adata.obs = tmp_adata.obs
